cambios

+++ src/aplicacion/control/PagoMensualDetallesController.java
@@ -791,7 +791,7 @@ public class PagoMensualDetallesController implements Initializable {
             rol.setDescripcion("Sueldo");
             rol.setIngreso(round(sueldoTotalTextValor));
             rol.setDias(diasTextValor);
-            rol.setHoras(normalesTextValor.intValue());
+            rol.setHoras(normalesTextValor);
             rol.setClave(Const.IP_SUELDO);
             pagoMesItems.add(rol);
         }
@@ -800,7 +800,7 @@ public class PagoMensualDetallesController implements Initializable {
             PagoMesItem rol = new PagoMesItem();
             rol.setDescripcion("Horas Extras");
             rol.setIngreso(extraTextValor);
-            rol.setHoras(suplementariasTextValor.intValue() + sobreTiempoTextValor.intValue());
+            rol.setHoras(suplementariasTextValor + sobreTiempoTextValor);
             rol.setClave(Const.IP_HORAS_EXTRAS);
             pagoMesItems.add(rol);
         }
