+++ src/aplicacion/control/PagosTotalEmpleadoController.java
@@ -839,7 +839,7 @@ public class PagosTotalEmpleadoController implements Initializable {
             rol.setDescripcion("Sueldo");
             rol.setIngreso(round(sueldoTotalTextValor));
             rol.setDias(diasTextValor);
-            rol.setHoras(normalesTextValor.intValue() );
+            rol.setHoras(normalesTextValor);
             rol.setClave(Const.IP_SUELDO);
             pagoMesItems.add(rol);
         }
@@ -848,7 +848,7 @@ public class PagosTotalEmpleadoController implements Initializable {
             PagoMesItem rol = new PagoMesItem();
             rol.setDescripcion("Horas Extras");
             rol.setIngreso(extraTextValor);
-            rol.setHoras((suplementariasTextValor.intValue()  + sobreTiempoTextValor.intValue() ));
+            rol.setHoras((suplementariasTextValor  + sobreTiempoTextValor));
             rol.setClave(Const.IP_HORAS_EXTRAS);
             pagoMesItems.add(rol);
         }
